Breakdown SPI to API, API Impl and SPI module
This is done mainly to ensure that there is no project dependency such
that it could not be included without intrusion. For example RESTful WS
client should not have SPI or its implementations but have the API and
may be the API Impl. Again Hub Common should be useable by the client as
well.

Signed-off-by: Imran M Yousuf <imyousuf@smartitengineering.com>