Add support for pagination of channels in API and SPI
For pagination support we needed the concept of channel index in storage
and thus added position to channel API. In SPI changes were necessary at
the time of conversion in the adapter. Addition of position caused changes
in udpate and delete in the test case. Before updating/deleting the channel
needs to be retrieved to get its position info.

Add pagination service to retrieve channels from persistent storage.

Signed-off-by: Imran M Yousuf <imyousuf@smartitengineering.com>