Add support for POST Event from HTML
The POST from HTML is checked by a Boolean variable isHtmlPost.
The message is formatted if it comes from HTML form. To make the
message human readable it is decoded to 'UTF-8'. After saving
the message a new line is added at the end to ensure that the
message is flushed to its listener.

Signed-off-by: kaisar <kaisar@smartitengineering.com>