Prepare the hub web application for HBase integration
The first step of the preparation was to ensure that context path in the
web.xml can be injected by maven just added maven-war-plugin's filtering
capability.
The second step is to be able to control certain dependencies, properties
and war file name from profile so that switching profile will generate
different SPI implementation configured wars.

In using Maven Build Profiles one must remember that if any pom is default
activated, such as the rdbms profile in this case, it will be deactivated
if any other profile in the POM is activated by condition or CLI, in our
case we will specify HBase.

Changed maven jetty plugin to use run-war goal so that it uses the actual
war to startup the webapp. We might consider adding some integration test
for this module at a later time.

Signed-off-by: Imran M Yousuf <imyousuf@smartitengineering.com>