Separate WAR modules for RDBMS and HBase version
This is done for ease of development and maintenance. Though the
modules are separated the only thing separate is their web.xml, but
they use the same JSPs, CSSs and JSs. So all changes should be
made to hub-rdbms (hub/) and it will reflect in hub-hbase.

The notable difference in usage is their artifact id, the old artifact id
is invalid, but it will be available in the snapshot repository for the time
being for backward portability.

Signed-off-by: Imran M Yousuf <imyousuf@smartitengineering.com>